# Conflicts:
#	src/main/java/me/vse/fintrackserver/configs/CorsConfig.java
#	src/main/java/me/vse/fintrackserver/controller/AccountController.java
#	src/main/java/me/vse/fintrackserver/controller/UserController.java
#	src/main/java/me/vse/fintrackserver/model/Group.java
#	src/main/java/me/vse/fintrackserver/services/AccountService.java
#	src/main/java/me/vse/fintrackserver/services/GroupService.java
#	src/main/resources/application-prod.properties
#	src/main/resources/liquibase/dbchangelog.xml

+++ src/main/java/me/vse/fintrackserver/services/CategoryService.java
@@ -4,11 +4,16 @@ import jakarta.persistence.EntityManager;
 import jakarta.transaction.Transactional;
 import me.vse.fintrackserver.enums.ErrorMessages;
 import me.vse.fintrackserver.mappers.CategoryMapper;
+import me.vse.fintrackserver.model.Account;
+import me.vse.fintrackserver.model.Asset;
 import me.vse.fintrackserver.model.Category;
+import me.vse.fintrackserver.model.User;
 import me.vse.fintrackserver.model.dto.CategoryDto;
 import me.vse.fintrackserver.repositories.CategoryRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.util.List;
 
 @Service
 public class CategoryService {
@@ -23,11 +28,30 @@ public class CategoryService {
     private CategoryRepository categoryRepository;
 
     @Transactional
+    public List<Category> getAll(String userId) {
+        if (userId == null) {
+            throw new IllegalArgumentException(ErrorMessages.USER_DOESNT_EXIST.name());
+        }
+
+        User user = entityManager.find(User.class, userId);
+        if (user == null) {
+            throw new IllegalArgumentException(ErrorMessages.USER_DOESNT_EXIST.name());
+        }
+
+        return user.getCategories();
+    }
+
+    @Transactional
     public Category add(CategoryDto categoryDto) {
+        User user = entityManager.find(User.class, categoryDto.getUserId());
+        if (user == null) {
+            throw new IllegalArgumentException(ErrorMessages.USER_DOESNT_EXIST.name());
+        }
+
         Category category = Category.builder()
                 .name(categoryDto.getName())
-                .color(categoryDto.getColor())
                 .icon(categoryDto.getIcon())
+                .user(user)
                 .build();
         entityManager.persist(category);
         return category;
@@ -43,13 +67,23 @@ public class CategoryService {
         if (category == null) {
             throw new IllegalArgumentException(ErrorMessages.CATEGORY_DOESNT_EXIST.name());
         }
+
+        User user = entityManager.find(User.class, categoryDto.getUserId());
+        if (user == null) {
+            throw new IllegalArgumentException(ErrorMessages.USER_DOESNT_EXIST.name());
+        }
+
+        if (!user.getCategories().contains(category)) {
+            throw new IllegalArgumentException(ErrorMessages.CATEGORY_DOESNT_EXIST.name());
+        }
+
         categoryMapper.updateCategoryMapperFromDto(categoryDto, category);
-        entityManager.persist(category);
+        categoryRepository.save(category);
         return category;
     }
 
     @Transactional
-    public void delete(String id) {
+    public void delete(String id, String userId) {
         if (id == null) {
             throw new IllegalArgumentException(ErrorMessages.CATEGORY_DOESNT_EXIST.name());
         }
@@ -57,6 +91,16 @@ public class CategoryService {
         if (category == null) {
             throw new IllegalArgumentException(ErrorMessages.CATEGORY_DOESNT_EXIST.name());
         }
+
+        User user = entityManager.find(User.class, userId);
+        if (user == null) {
+            throw new IllegalArgumentException(ErrorMessages.USER_DOESNT_EXIST.name());
+        }
+
+        if (!user.getCategories().contains(category)) {
+            throw new IllegalArgumentException(ErrorMessages.CATEGORY_DOESNT_EXIST.name());
+        }
+
         categoryRepository.delete(category);
     }
 }
